Add basic clock edit mode
The ability is still very limited.  Enter clockedit by the cle
command.  Currently, only the current day can be edited.  The clocks
of the day are pre loaded and can be viewed with the ls command.
The commands start and end allow to override the start or end time of
a clock.   The syntax is:

start (i) (HH:MM:SS)
end (i) (HH:MM:SS)

'i' is the index of the clock.  The index is the first number shown
by the ls command.  While annoying, seconds are required for now.

This also panics if the time format is wrong.

Exiting the clockedit mode with 'apply' command will override the clocks
in the document but exiting with 'cancel' will discard the changes.