Use absraction for CLI
The command line interface (CLI) is not abstracted.  A callback trait
CliCallbacks implements the functions to print and read text.

This sould easily allow to support different kind of architectures.  For
example could it be used for a more advanced UI which contains the CLI
a separate section for faster access.  Another use case would be that the
application can act as a server and sends and receives command through
sockets.  Finally this should be an important step to run the application
in a browser.